Split tests in different packages so that they can run in parallel
