+++ tools/integration-tests/runTests.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+TEST_NAMES='autopeering common drng message'
+
 echo "Build GoShimmer image"
 docker build -t iotaledger/goshimmer ../../.
 
@@ -9,10 +11,12 @@ docker pull gaiaadm/pumba:latest
 docker pull gaiadocker/iproute2:latest
 
 echo "Run integration tests"
-docker-compose -f tester/docker-compose.yml up --abort-on-container-exit --exit-code-from tester --build
 
-echo "Create logs from containers in network"
-docker logs tester &> logs/tester.log
+for name in $TEST_NAMES
+do
+  TEST_NAME=$name docker-compose -f tester/docker-compose.yml up --abort-on-container-exit --exit-code-from tester --build
+  docker logs tester &> logs/"$name"_tester.log
+done
 
 echo "Clean up"
 docker-compose -f tester/docker-compose.yml down
