+++ tools/integration-tests/tester/docker-compose.yml
@@ -5,7 +5,7 @@ services:
     container_name: tester
     image: golang:1.14
     working_dir: /tmp/goshimmer/tools/integration-tests/tester
-    entrypoint: go test ./tests -v -mod=readonly -timeout 30m
+    entrypoint: go test ./tests/${TEST_NAME} -v -mod=readonly -timeout 30m
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock:ro
       - ../../..:/tmp/goshimmer:ro
