+++ tools/integration-tests/tester/tests/autopeering/main_test.go
@@ -1,9 +1,4 @@
-// Package tests provides the possibility to write integration tests in regular Go style.
-// The integration test framework is initialized before any test in the package runs and
-// thus can readily be used to create networks.
-//
-// Each tested feature should reside in its own test file and define tests cases and networks as necessary.
-package tests
+package autopeering
 
 import (
 	"os"
